Auth

+++ controllers/controller.js
@@ -1,13 +1,13 @@
 const { ToDo } = require('../models/index.js')
 class Controller {
-    static read(req, res) {
+    static read(req, res, next) {
         // console.log('masuk')
         console.log(req.decoded)
-        ToDo.findAll()
+        ToDo.findAll({where: {UserId: req.decoded.id}})
             .then(todos => res.status(200).json(todos))
-            .catch(err => res.status(500).json(err))
+            .catch(err => next(err))
     }
-    static create(req, res) {
+    static create(req, res, next) {
         // res.status(200).json(req.body)
         let newItem = {
             title: req.body.title,
@@ -19,22 +19,24 @@ class Controller {
         ToDo.create(newItem)
             .then(todo => res.status(201).json(todo))
             .catch(err => {
-                if(err.name === 'SequelizeValidationError') return res.status(400).json(err)
-                else return res.status(500).json(err)
+                return next(err)
             })
         // res.status(200).json(newItem)
         // ToDo.create()
     }
-    static readId(req, res) {
+    static readId(req, res, next) {
         let id = req.params.id;
-        ToDo.findAll({where: {id: id}})
+        let UserId = req.decoded.id;
+        console.log(req.decoded)
+        ToDo.findOne({where: {id: id, UserId: UserId}})
             .then(todos => {
-                if(todos.length < 1) return res.status(404).json({error: 'Not Found!'})
+                if(todo.UserId !== req.decoded.id) throw {status: 400, customName: 'You do not have access to this item.'}
+                else if(todos.length < 1) throw {status: 404, customName: 'Item not found!'}
                 else return res.status(200).json(todos)}
                 )
-            .catch(err => res.status(500).json(err))
+            .catch(err => next(err))
     }
-    static update(req, res) {
+    static update(req, res, next) {
         let id = req.params.id;
         // console.log(req.body)
         let updatedItem = {
@@ -43,27 +45,37 @@ class Controller {
             status: req.body.status,
             due_date: req.body.due_date
         }
-        ToDo.update(updatedItem, {where: {id: id}})
+        ToDo.findOne({where: {id: id}})
             .then(todo => {
-                if(todo[0] == 0) return res.status(404).json({error: 'Not found!'})
-                else return ToDo.findAll({where: {id: id}})
+                if(todo.UserId == req.decoded.id) return ToDo.update(updatedItem, {where: {id: id}})
+                else throw {status: 400, customName: 'You do not have access to this item.'}
+            })
+            .then(todo => {
+                if(todo[0] == 0) throw {status: 404, customName: 'Item not found!'}
+                else return ToDo.findOne({where: {id: id}})
             })
             .then(todo => {
                 return res.status(201).json(todo)
             })
             .catch(err => {
-                if(err.name === 'SequelizeValidationError') return res.status(400).json(err)
-                else return res.status(500).json(err)
+                return next(err)
             })
     }
     static delete(req, res) {
         let id = req.params.id;
-        ToDo.destroy({where: {id: id}})
+        ToDo.findOne({where: {id: id}})
             .then(todo => {
-                if(todo == 0) return res.status(404).json({error: 'Not found!'})
-                else return res.status(200).json(todo)
+                if(todo.UserId == req.decoded.id) return ToDo.destroy({where: {id: id}})
+                else throw {status: 400, customName: 'You do not have access to this item.'}
             })
-            .catch(err => res.status(500).json(err))
+            .then(todo => {
+                if(todo[0] == 0) throw {status: 404, customName: 'Item not found!'}
+                else return ToDo.findOne({where: {id: id}})
+            })
+            .then(todo => {
+                return res.status(201).json(todo)
+            })
+            .catch(err => next(err))
     }
 }
